requirements.txt
absl-py==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==21.4.0
backcall 
beautifulsoup4
bleach==5.0.0
blis==0.4.1
brotlipy==0.7.0
cachetools==4.2.4
catalogue==1.0.2
certifi==2021.10.8
cffi 
chardet==3.0.4
click==8.1.7
# conda==4.10.3
# conda-build==3.21.5
# conda-package-handling
cryptography 
cycler==0.11.0
cymem==2.0.8
dataclasses==0.6
debugpy==1.6.0
decorator 
defusedxml==0.7.1
dnspython==2.1.0
docopt==0.6.2
download==0.3.5
easydict==1.10
entrypoints==0.4
fair-esm==2.0.0
fastjsonschema==2.15.3
filelock 
fire==0.4.0
fonttools==4.38.0
future==0.18.3
glob2 
google-auth-oauthlib==0.4.6
grpcio==1.59.3
idna 
imageio==2.25.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython 
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi 
Jinja2==3.1.1
joblib==1.3.2
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab==2.2.5
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
libarchive-c 
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib
matplotlib-inline
# mindspore-dev 
# MinkowskiEngine==0.5.4
mistune==0.8.4
# mkl-fft==1.3.1
# mkl-random
# mkl-service==2.4.0
# MSAdapter==0.0.1a0
six
murmurhash==1.0.10
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
ninja==1.11.1
nltk==3.2.5
notebook==6.4.11
numpy==1.19.4
oauthlib==3.2.2
olefile==0.46
open3d-python
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso 
pexpect 
pickleshare 
Pillow==8.4.0
pkginfo==1.7.1
plac==1.1.3
pqi==3.0.0
preshed==3.0.9
prometheus-client==0.14.1
prompt-toolkit 
protobuf==3.20.3
psutil 
ptyprocess 
pyaml==21.10.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat==0.6.3
pycparser 
Pygments 
pyOpenSSL 
pyparsing==3.0.8
pypng==0.20220715.0
pyrsistent==0.18.1
PySocks 
python-dateutil==2.8.2
python-etcd==0.4.5
pytorch-metric-learning==2.3.0
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.3.1
pyzmq==22.3.0
regex==2023.10.3
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.9
# ruamel-yaml-conda
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
soupsieve 
spacy==2.2.4
srsly==1.0.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
terminado==0.13.3
thinc==7.4.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tokenizers==0.10.3
# torch==1.12.1+cu102
# torchac==0.9.3
# torchaudio==0.12.1+cu102
# torchelastic==0.2.0
# torchtext==0.3.1
# torchvision==0.13.1+cu102
# tornado==6.1
tqdm==4.53.0
traitlets 
transformers==4.3.2
typing-extensions==4.5.0
urllib3==1.25.11
wasabi==0.10.1
wcwidth 
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
xlrd==1.2.0
zipp==3.8.0
